% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_palettes.R
\docType{data}
\name{exp_palettes}
\alias{exp_palettes}
\title{Example list of palettes}
\format{An object of class \code{list} of length 14.}
\usage{
exp_palettes
}
\description{
Use \code{\link{img2hex}} to construct palettes of desired length.
}
\keyword{datasets}
